+++ payment-sdk/src/main/java/com/worldline/sips/configuration/PaymentConfiguration.java
@@ -7,7 +7,7 @@ public class PaymentConfiguration {
     /**
      * The targeted version of the API
      */
-    public static final String INTERFACE_VERSION = "IR_WS_2.35";
+    public static final String INTERFACE_VERSION = "IR_WS_2.46";
 
     private PaymentConfiguration() {
         // Nothing to see here