+++ sdk-common/src/main/java/com/worldline/sips/model/NamedPaymentMeanType.java
@@ -1,0 +1,11 @@
+package com.worldline.sips.model;
+
+public enum NamedPaymentMeanType implements PaymentMeanType {
+    CARD, CREDIT_TRANSFER, DIRECT_DEBIT,
+    VOUCHER, WALLET, ONLINE_CREDIT, EMPTY, PROVIDER;
+
+    @Override
+    public String getName() {
+        return name();
+    }
+}