+++ sdk-common/src/main/java/com/worldline/sips/model/PaymentMeanType.java
@@ -3,16 +3,19 @@ package com.worldline.sips.model;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import org.apache.commons.lang3.StringUtils;
 
-public enum PaymentMeanType {
-    CARD, CREDIT_TRANSFER, DIRECT_DEBIT,
-    VOUCHER, WALLET, ONLINE_CREDIT, EMPTY, PROVIDER;
-
+public interface PaymentMeanType {
+    String getName();
+  
+  
     @JsonCreator
-    public static PaymentMeanType fromValue(String value) {
+    static PaymentMeanType fromValue(String value) {
         if (StringUtils.isBlank(value)) {
-            return PaymentMeanType.EMPTY;
+            return NamedPaymentMeanType.EMPTY;
         }
-        return PaymentMeanType.valueOf(value);
+        try {
+            return NamedPaymentMeanType.valueOf(value);
+        } catch (IllegalArgumentException e) {
+            return new UnknownPaymentMeanType(value);
+        }
     }
-
 }