+++ sdk-common/src/main/java/com/worldline/sips/model/UnknownPaymentMeanBrand.java
@@ -1,0 +1,14 @@
+package com.worldline.sips.model;
+
+public class UnknownPaymentMeanBrand implements PaymentMeanBrand {
+    private final String realName;
+  
+    public UnknownPaymentMeanBrand(String realName) {
+        this.realName = realName;
+    }
+  
+    @Override
+    public String getRealName() {
+        return realName;
+    }
+}