+++ sdk-common/src/main/java/com/worldline/sips/model/UnknownPaymentMeanType.java
@@ -1,0 +1,14 @@
+package com.worldline.sips.model;
+
+public class UnknownPaymentMeanType implements PaymentMeanType {
+    private final String name;
+  
+    public UnknownPaymentMeanType(String name) {
+        this.name = name;
+    }
+  
+    @Override
+    public String getName() {
+        return name;
+    }
+}